Adding common kick drum strategy

+++ src/generators/kick-drum.generator.ts
@@ -12,8 +12,6 @@ export class KickDrumGenerator {
 
     async generate() {
         var track = await AbletonJs.createMidiTrackIfNotExists(KickDrumGenerator.TRACK_NAME);
-        await this.generateFourOnFloor(track);
-        await this.generateDancehall(track);
         for (var i = 0; i < Config.KickDrum.strategies.length; i++) {
             var strategy = Config.KickDrum.strategies[i];
             await this.generatePhraseFromStrategy(track, strategy);
@@ -21,19 +19,7 @@ export class KickDrumGenerator {
         return track;
     }
 
-    async generateFourOnFloor(track: Track) {
-        var phrase = this.generatePhraseFromPattern("x---x---x---x---", "FourOnFloor").double();
-        await AbletonJs.insertMidiClip(track, phrase.toMidiClip());
-    }
-
-    async generateDancehall(track: Track) {
-        var phrase = this.generatePhraseFromPattern("x--x--x-x--x--x-", "Dancehall1").double();
-        await AbletonJs.insertMidiClip(track, phrase.toMidiClip());
-
-        phrase = this.generatePhraseFromPattern("x--x----x--x----", "Dancehall2 ").double();
-        await AbletonJs.insertMidiClip(track, phrase.toMidiClip());
-    }
-
+   
     async generatePhraseFromStrategy(track: Track, strategy: IClipGenerationStrategy) {
         for(var i = 0; i < strategy.numberOfClips; i++) {
             var phrase = strategy.generate();
@@ -41,28 +27,12 @@ export class KickDrumGenerator {
         }
     }
 
-    generatePhraseFromPattern(pattern: string, name: string = null) {
-        if (pattern.length % 4 !== 0) { throw new Error("Must be 4/4 time yo"); }
-        var sixteenthNotes = pattern.length;
-        var beats = sixteenthNotes / 4;
-        var bars = beats / 4;
-        var phrase = new Phrase(bars, name);
-        for (var i = 0; i < sixteenthNotes; i++) {
-            var action = pattern[i];
-            if (action === "x") {
-                var note = Note.fromNoteName("C3", i / 4, NoteDuration.Sixteenth);
-                phrase.addNote(note);
-            }
-
-        }
-
-        return phrase;
-    }  
-
     async clearClips(): Promise<void> {
 
         var track = await AbletonJs.getTrackByName(KickDrumGenerator.TRACK_NAME);
-        await AbletonJs.deleteAllMidiClips(track);
+        if(track){
+            await AbletonJs.deleteAllMidiClips(track);
+        }
 
         return;
     }
